Create Simulate_MRI_data.r
R script file to simulate ventricle and sum of the left and right hippocampus volume data. The posterior means of the Bayesian linear mixed effects model, expression (3) in the manuscript, were used to simulate the longitudinal data for N = 260 participants across 4 time points.
To replicate the simulated data as close as possible to the original AIBL data used in the manuscript, we incorporate an unbalanced structure (decreasing follow-up observations as time progresses) and participants at various levels of conversion (HC to MCI, MCI to AD and HC to AD).